// frontend/src/app/core/api/models/studentGradeRowDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StudentGradeRowDto { 
    studentId?: string;
    studentName?: string;
    group?: string;
    finalGradeId?: string;
    grade?: number;
    provisional?: boolean;
    completionDate?: string;
}